requirements.txt
fastapi==0.95.2
pydantic==1.10.8
uvicorn==0.22.0
python-multipart==0.0.6
python-dotenv==1.0.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]
sqlalchemy==2.0.15
psycopg2-binary==2.9.6
jinja2==3.1.2
pytest==7.3.1
httpx==0.24.1